% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpower.R
\name{estimate_p123_data}
\alias{estimate_p123_data}
\title{Estimate p1, p2, p3 from pilot data}
\usage{
estimate_p123_data(x, y)
}
\arguments{
\item{x}{Numeric Vector. Control group observations}

\item{y}{Numeric vector. Treatment group observations}
}
\value{
A list with p1, p2, p3
}
\description{
Given observed data from two groups, estimates p1, p2, p3 using
pairwise comparisons. This follows estimators described on page 363 of
T=Book
}
