% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Power_Curve.R
\name{plot_power_curve}
\alias{plot_power_curve}
\title{Plot power curves across a range of effect sizes}
\usage{
plot_power_curve(
  theta_range,
  alpha = 0.05,
  power = 0.8,
  k = 1,
  rdist = rnorm,
  alternative = "two.sided",
  exact = FALSE,
  correct = FALSE,
  nsim_est = 1e+05,
  nsim_pow = 5000
)
}
\description{
Plot power curves across a range of effect sizes
}
\examples{
results <- plot_power_curve(theta_range = seq(0.3, 1.2, by = 0.1))
results2 <- plot_power_curve(theta_range = c(0.4, 0.6, 0.8, 1.0, 1.2),nsim_pow=3000)

}
