% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulates_ttest_mean.R
\name{simulate_two_sample}
\alias{simulate_two_sample}
\title{Simulates Two Sample t-test for Power Estimation}
\usage{
simulate_two_sample(n1, n2, margin, sigma, alpha = 0.05, n_sim = 10000)
}
\arguments{
\item{n1}{sample size in group 1}

\item{n2}{Sample size in group 2}

\item{margin}{(µ2 - µ1)}

\item{sigma}{SD}

\item{alpha}{Significance level}

\item{n_sum}{Number of simulations}
}
\value{
Estimated Power
}
\description{
Simulates Two Sample t-test for Power Estimation
}
