% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpower.R
\name{wilcox_power}
\alias{wilcox_power}
\title{Simulate power of the Wilcoxon rank-sum test}
\usage{
wilcox_power(
  n1,
  n2,
  theta,
  alpha = 0.05,
  rdist = rnorm,
  nsim = 2000,
  alternative = "two.sided",
  exact = FALSE,
  correct = FALSE
)
}
\arguments{
\item{n1}{Integer. Sample size for group 1 (control)}

\item{n2}{Integer. Sample size for group 2 (treatment)}

\item{theta}{Numeric. Location Shift}

\item{alpha}{Numeric. Significance level Default: 0.05}

\item{rdist}{Function. Random generator. Default: rnorm}

\item{nsim}{Integer. Number of simulation replicates. Default is 2000}

\item{alternative}{Character. Type of alternative hypothesis;
one of "two.sided", "less", "greater". Default: two.sided}

\item{exact}{logical Computes exact p-value. Default: FALSE}

\item{correct}{Logical. Whether to apply continuity correction.
Default: FALSE}
}
\value{
Numeric. Estimated Power.
}
\description{
Estimates the empirical power of \code{\link[stats]{wilcox.test}}
under a location shift model by Monte Carlo simulation.
}
\examples{
wilcox_power(n1 = 20, n2 = 20, theta = 0.7, rdist = rnorm, nsim = 1000)
}
