% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpower.R
\name{wpower}
\alias{wpower}
\title{Compares Noether and TrialSize sample size calculations for Wilcoxon
rank sum test.}
\usage{
wpower(
  theta,
  alpha = 0.05,
  power = 0.8,
  k = 1,
  p1 = NULL,
  p2 = NULL,
  p3 = NULL,
  rdist = rnorm,
  pilot_x = NULL,
  pilot_y = NULL,
  alternative = "two.sided",
  exact = FALSE,
  correct = FALSE,
  nsim_est = 1e+05,
  nsim_pow = 5000,
  ...
)
}
\arguments{
\item{theta}{Numeric. Location shift (treatment effect)}

\item{alpha}{Numeric. Significance Level. Default is 0.05}

\item{power}{Numeric. Target power. Default is 0.8}

\item{k}{Numeric. Allocation ratio \eqn{k=n_1/n_2}. Default is 1

If supplied, used as effect size inputs for the TrialSize calculations.}

\item{p1}{Optional numeric. P(X>Y)}

\item{p2}{Optional numeric. P(Y >= X_1 AND Y >= X_2)}

\item{p3}{Optional numeric. P(Y_1 >= X AND Y_2 >= X)}

\item{rdist}{Function. Random generator. Default rnorm}

\item{pilot_x}{Optional numeric vectors. If supplied, used to
estimate \eqn{p_1,p_2,p_3} from pilot data}

\item{pilot_y}{Optional numeric vectors. If supplied, used to
estimate \eqn{p_1,p_2,p_3} from pilot data}

\item{alternative}{Character. Passed to \code{\link[stats]{wilcox.test}}.}

\item{exact}{Logical Passed to \code{\link[stats]{wilcox.test}}.}

\item{correct}{Logical. Passed to \code{\link[stats]{wilcox.test}}.}

\item{nsim_est}{Integer Number of simulations for estimating \eqn{p_1,p_2,p_3}}

\item{nsim_pow}{Integer Number of simulations for estimating power}

\item{...}{Additional arguments passed to \code{\link[stats]{wilcox.test}}.}
}
\value{
An object of class \code{wpower}: a list with components
\describe{
\item{trialsize}{list with \code{n1}, \code{n2}, \code{N_total}, and simulated \code{power}.}
\item{noether}{list with \code{n1}, \code{n2}, \code{N_total}, and simulated \code{power}.}
\item{theta, alpha, target_power, k}{Input Settings}
\item{rdist}{Generator function used for simulation}
\item{alternative, exact, correct}{Wilcoxon test options used in simulation }
}
}
\description{
Computes sample sizes using two approaches for the two sample Wilcoxon rank
sum test
}
\details{
The function then estimates achieved power for both designs by simulation
under the specified location shift \code{theta} and distribution \code{rdist}.
}
\examples{
# Example 1: Simulated estimates of p1, p2, p3 under a normal shift model
Example1 <- wpower(theta = 0.7, rdist = rnorm, nsim_est = 50000, nsim_pow = 1000)
Example1$trialsize
Example1$noether

}
