% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClinSim-package.R
\docType{package}
\name{ClinSim-package}
\alias{ClinSim}
\alias{ClinSim-package}
\title{ClinSim:}
\description{
Describe
}
\details{
The main functions are

The following example datasets are provided.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
