% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpower.R
\name{estimate_p123}
\alias{estimate_p123}
\title{Estimate p1, p2, and p3 from a location shift model by simulation}
\usage{
estimate_p123(theta, rdist, nsim = 30000)
}
\arguments{
\item{theta}{Numeric Treatment Effect. Location Shift}

\item{rdist}{Function. Random generator for the error distribution. The default is \code{rnorm}}

\item{nsim}{Number of simulation}
}
\value{
A list with p1, p2, p3
}
\description{
Simulates independent samples from a chosen distribution under the model
\eqn{X} and \eqn{X'} are i.i.d., and \eqn{Y = X' + \theta}.
}
\details{
Estimates:

\eqn{p_1 = P(Y >= X)}
\eqn{p_2 = P(Y >= X_1 \;&\; Y >= X_2)}
\eqn{p_3 = P(Y_1 >= X \;&\; Y_2 >= X)}
}
\examples{
estimate_p123(theta = 0.7, rdist = rnorm)
estimate_p123(theta = 0.7, rdist= rlogis, nsim=5000)
}
