% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpower.R
\name{estimate_p123}
\alias{estimate_p123}
\title{Estimate p1, p2, p3 from a location shift model via simulation}
\usage{
estimate_p123(theta, rdist, nsim = 30000)
}
\arguments{
\item{theta}{Numeric Treatment Effect. Location Shift}

\item{rdist}{Function. Default: rnorm}

\item{nsim}{Number of simulation}
}
\value{
A list with p1, p2, p3
}
\description{
p1 = P(Y >= X)
p2 = P(Y >= X1 AND Y >= X2)
p3 = P(Y1 >= X AND Y2 >= X)
}
