% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpower.R
\name{noether_N}
\alias{noether_N}
\title{Noether (1987) sample size formula}
\usage{
noether_N(alpha, beta, p1, k = 1)
}
\arguments{
\item{alpha}{significance level}

\item{beta}{Type II error (1-Power)}
}
\value{
Total Sample Size N
}
\description{
simpler only needs p1
}
