% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{print.wpower}
\alias{print.wpower}
\title{Print method for wpower object}
\usage{
\method{print}{wpower}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{wpower()}}

\item{...}{Additional arguments.}
}
\value{
Returns the input object \code{x}
}
\description{
Displays a summary of a \code{wpower} object, including the effect size
used, the calculated sample size, the simulated power and the method used.
}
