% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{summary.wpower}
\alias{summary.wpower}
\title{Summary method for wpower object}
\usage{
\method{summary}{wpower}(object, ...)
}
\arguments{
\item{object}{A wpower object}

\item{...}{additional arguments}
}
\value{
Returns the input object \code{object}
}
\description{
Detailed summary Method with explanation of results and estimated
\eqn{p_1,p_2,p_3}
}
