% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{summary.wpower}
\alias{summary.wpower}
\title{Summary method for}
\usage{
\method{summary}{wpower}(object, ...)
}
\arguments{
\item{object}{A wpower object}

\item{...}{additional arguments}
}
\description{
Detailed summary Method with explanation of results
}
