% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpower.R
\name{wilcox_power}
\alias{wilcox_power}
\title{Simulate power of the Wilcoxon rank-sum test}
\usage{
wilcox_power(
  n1,
  n2,
  theta,
  alpha = 0.05,
  rdist = rnorm,
  nsim = 2000,
  alternative = "two.sided",
  exact = FALSE,
  correct = FALSE
)
}
\arguments{
\item{n1}{Integer. Sample size for group 1}

\item{n2}{Integer. Sample size for group 2}

\item{alpha}{Numeric Significance level Default: 0.05}

\item{rdist}{Default: rnorm}

\item{nsim}{Integer. Number of simulation replicates Default}

\item{alternative}{Character Type of alternative hypothesis
one if "two.sided", "less", "greater". Default: two.sided}

\item{exact}{logical Compuyer exact p-value. Default: FALSE}

\item{correct}{Logical. Whether to apply continuity correction.
Default: FALSE}

\item{thera}{Numeric. Location Shift}
}
\value{
Numeric, Estimated Power
}
\description{
Simulate power of the Wilcoxon rank-sum test
}
